Revert "lgtm handler: find lgtm comment and add label"
