+++ mungegithub/README.md
@@ -54,8 +54,7 @@ A small amount of information about some of the individual mungers inside each o
 * cherrypick-label-unapproved - adds `do-not-merge` label to PRs against a release-\* branch which do not have `cherrypick-approved`
 * comment-deleter - deletes comments created by the k8s-merge-robot which are no longer relevant. Such as comments about a rebase being required if it has been rebased.
 * comment-deleter-jenkins - deleted comments create by the k8s-bot jenkins bot which are no longer relevant. Such as old test results.
-* lgtm-handler - adds `lgtm` label if reviewer has commented "LGTM"; or notifies reviewer if non-reviewer has commented "LGTM";
-or removes `lgtm` label if a PR is changed after the label was added
+* lgtm-after-commit - removes `lgtm` label if a PR is changed after the label was added
 * needs-rebase - adds and removes a `needs-rebase` label if a PR needs to be rebased before it can be applied.
 * path-label - adds labels, such as `kind/new-api` based on if ANY file which matches changed
 * rebuild-request - Looks for requests to retest a PR but which did not reference a flake.
